feat: upgrade example, but still issue
