--- example/.ruby-version
@@ -1,1 +1,0 @@
-2.7.6
