+++ example/android/build.gradle
@@ -1,54 +1,21 @@
-import org.apache.tools.ant.taskdefs.condition.Os
-
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-  ext {
-    buildToolsVersion = "31.0.0"
-    minSdkVersion = 21
-    compileSdkVersion = 31
-    targetSdkVersion = 31
+    ext {
+        buildToolsVersion = "33.0.0"
+        minSdkVersion = 21
+        compileSdkVersion = 33
+        targetSdkVersion = 33
 
-    if (System.properties['os.arch'] == "aarch64") {
-      // For M1 Users we need to use the NDK 24 which added support for aarch64
-      ndkVersion = "24.0.8215888"
-    } else {
-      // Otherwise we default to the side-by-side NDK version from AGP.
-      ndkVersion = "21.4.7075529"
+        // We use NDK 23 which has both M1 support and is the side-by-side NDK version from AGP.
+        ndkVersion = "23.1.7779620"
     }
-    kotlinVersion = '1.6.0'
-  }
-  repositories {
-    google()
-    mavenCentral()
-  }
-  dependencies {
-    classpath("com.android.tools.build:gradle:7.0.4")
-    classpath("com.facebook.react:react-native-gradle-plugin")
-    classpath("de.undercouch:gradle-download-task:4.1.2")
-    // NOTE: Do not place your application dependencies here; they belong
-    // in the individual module build.gradle files
-  }
-}
-
-allprojects {
-  repositories {
-    maven {
-      // All of React Native (JS, Obj-C sources, Android binaries) is installed from npm
-      url("$rootDir/../node_modules/react-native/android")
+    repositories {
+        google()
+        mavenCentral()
     }
-    maven {
-      // Android JSC is installed from npm
-      url("$rootDir/../node_modules/jsc-android/dist")
+    dependencies {
+        classpath("com.android.tools.build:gradle:7.3.1")
+        classpath("com.facebook.react:react-native-gradle-plugin")
     }
-    mavenCentral {
-      // We don't want to fetch react-native from Maven Central as there are
-      // older versions over there.
-      content {
-        excludeGroup "com.facebook.react"
-      }
-    }
-    google()
-    maven { url 'https://www.jitpack.io' }
-  }
 }
