+++ example/app.json
@@ -1,0 +1,4 @@
+{
+  "name": "TfliteExample",
+  "displayName": "TfliteExample"
+}
