+++ example/index.js
@@ -1,4 +1,5 @@
 import { AppRegistry } from 'react-native';
 import App from './src/App';
+import { name as appName } from './app.json';
 
-AppRegistry.registerComponent('main', () => App);
+AppRegistry.registerComponent(appName, () => App);
