+++ example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-native-tflite-example",
+  "name": "TfliteExample",
   "description": "Example app for react-native-tflite",
   "version": "0.0.1",
   "private": true,
@@ -10,19 +10,18 @@
     "pods": "pod-install --quiet"
   },
   "dependencies": {
-    "react": "17.0.2",
-    "react-native": "0.68.2",
+    "react": "18.2.0",
+    "react-native": "0.71.7",
     "react-native-image-picker": "^4.8.4",
     "react-native-permissions": "^3.6.0",
-    "react-native-reanimated": "^2.9.1",
-    "react-native-vision-camera": "^2.14.0"
+    "react-native-reanimated": "^2.14.4",
+    "react-native-vision-camera": "^2.15.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/runtime": "^7.12.5",
-    "babel-plugin-module-resolver": "^4.1.0",
-    "metro-react-native-babel-preset": "^0.67.0",
-    "patch-package": "^6.4.7",
-    "postinstall-postinstall": "^2.1.0"
+    "@babel/core": "^7.20.0",
+    "@babel/preset-env": "^7.20.0",
+    "@babel/runtime": "^7.20.0",
+    "metro-react-native-babel-preset": "0.73.9",
+    "babel-plugin-module-resolver": "^4.1.0"
   }
 }
